pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raise"
version = "0.1.0"
description = "RAG hyper-parameter search and evaluation toolkit"
readme = "readme.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "tqdm>=4.64",
  "torch>=2.0",
  "torchvision>=0.15",
  "transformers>=4.35",
  "sentence-transformers>=2.2",
  "chromadb>=1.5.0",
  "bert-score>=0.3",
  "nltk>=3.8",
  "pillow>=9.0",
  "datasets>=2.14",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
raise-eval = "raisex.cli.eval_cli:main"
raise-algo = "raisex.cli.algo_cli:main"
